Refine stack op to improve xlnet performance, test=develop
stack's wait cost a lot of cpu time, use cuda kernel to do memory copy
will reduce cpu time.
